+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0]
+
+### Added
+
+- Add `InjectorBuilder` to programmatically configure the injector
+- Add `ContainerConfig` to apply container aliasing for injector
+- Add `ServiceConfig` to apply service definitions to injector
+
 ## [1.0.2]
 
 ### Fixed